Fix possible narrowing issue with implicit promotion to double.
Former-commit-id: fbe7db01398ad5af05736d0d6fd88b5869c34e28

+++ analysis/HistogramAccumulator.cc
@@ -253,6 +253,6 @@ void HistogramAccumulator::AccumulateSingleValue(const double&  oldMean,
 						 double&        newMean,
 						 double&        newVari) const
 {
-  newMean = oldMean + ((x - oldMean) / nEntriesAccumulated);
+  newMean = oldMean + ((x - oldMean) / (double)nEntriesAccumulated);
   newVari = oldVari + ((x - oldMean) * (x - newMean));
 }
